Fix all "conversion from ‘qsizetype’" compiler warnings
Variable, parameter and return types were changed to qsizetype when this was
trivial, required minor additional changes or the type should logically be
qsizetype (e.g. for indixes).
In all other cases qsizetype is wrapped in static_cast<int>(...).

+++ app/FileSystem/sizecalculator.cpp
@@ -173,5 +173,5 @@ void SizeCalculator::done()
 
 void SizeCalculator::emitProgress()
 {
-    emit progressChanged(((m_urls.length() - m_nextUrls.length()) * 100) / m_urls.length());
+    emit progressChanged(static_cast<int>(((m_urls.length() - m_nextUrls.length()) * 100) / m_urls.length()));
 }
